requirements.txt
boto3~=1.34.108
PyYAML~=6.0.1
botocore~=1.34.108
requests~=2.32.3